Event copies all images to event bucket, don't try to delete buckets if offline.

+++ spec/models/image_service_spec.rb
@@ -5,6 +5,7 @@ RSpec.describe ImageService, :type => :model do
   let(:bucket_name) { 'test_palantir' }
 
   after(:all) do
+    return unless online?
     ImageService.new.delete_bucket('test_palantir')
   end
